"checked in attendee list"
